add migrations in db

+++ api/src/main/java/org/genboard/model/Coord.java
@@ -2,31 +2,35 @@ package org.genboard.model;
 
 public class Coord {
 	
-	private int x;
+	private Double x;
 	
-	private int y;
+	private Double y;
 
-	
 	
 	public Coord() {
 		super();
 	}
 
-	public int getX() {
+
+	public Double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+
+	public void setX(Double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+
+	public Double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+
+	public void setY(Double y) {
 		this.y = y;
 	}
+
